解决新增时springmvc封装表单数据时String到float出错，在实体类中类型都写成包装类型 Float、Integer

+++ src/main/java/com/qijun/goods/goodstype/mapper/GoodsCommonMapper.java
@@ -10,7 +10,7 @@ import java.util.List;
  */
 public interface GoodsCommonMapper {
 
-    int insert(GoodsCommon goodsCommon);
+    long insert(GoodsCommon goodsCommon);
 
     int update(GoodsCommon goodsCommon);
